% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_drivers.R
\name{assign_drivers}
\alias{assign_drivers}
\title{Assign driver mutations in a general karyotype}
\usage{
assign_drivers(x, rho = 0.01)
}
\arguments{
\item{x}{A mobster hierarchical fit object}

\item{rho}{overdispersion parameter to use in the beta binomial}
}
\description{
Assign driver mutations in a general karyotype
}
\examples{
data()

}
