% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mutation_posteriori.R
\name{assign_mutations_posteriori}
\alias{assign_mutations_posteriori}
\title{Assign mutation a posteriori}
\usage{
assign_mutations_posteriori(x)
}
\arguments{
\item{x}{mobsterh fit object}
}
\value{
a mobsterh object with new cluster annotations and a new field posteriori_measures with cluster probabilities
}
\description{
A function to assign filtered and non-subsetted mutations in used karyotypes
}
\examples{
library(mobster)
library(dplyr)
assign_mutations_posteriori(mobster::fit_example_mobsterh$best)$data \%>\% filter(posteriori_annot) \%>\% head(5)
}
