% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_statistics.R
\name{bootstrapped_statistics}
\alias{bootstrapped_statistics}
\title{Compute boostrap statistics from a bootstrap run.}
\usage{
bootstrapped_statistics(x, bootstrap_results,
  bootstrap = "nonparametric", alpha = 0.025)
}
\arguments{
\item{x}{A MOBSTER fit.}

\item{bootstrap_results}{The results of running function \code{mobster_bootstrap}}

\item{bootstrap}{Type of boostrap: \code{"parametric"} or \code{"nonparametric"}}

\item{alpha}{Alpha-level empirical CIs for the bootstrap distribution.}
}
\value{
A list with the bootstrapped values, the model frequency, the bootstraped statistics and
the co-clustering probability (only for nonparametric boostraps).
}
\description{
From a bootstrap run several statistics are computed.
These include the model frequency, the statistics of each fit parameter
and, if the bootstrap is nonparametric, also the co-clustering probability
of the input points. All the bootstrapped values are also returned. For the
bootstrapped statistics the confidence intervals are returned given a 
required input alpha value (default 0.025).
}
