% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_genes_dnds}
\alias{cancer_genes_dnds}
\title{List of cancer genes to compute dnds values.}
\format{List of cancer genes to compute dnds values.}
\usage{
data(cancer_genes_dnds)
}
\description{
List of gene ids (HuGO format) developed from Martincorena et al. to compute dnds values.
}
\examples{
data(cancer_genes_dnds)
print(cancer_genes_dnds$best)
}
\keyword{datasets}
