% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_getters.R
\name{Clusters}
\alias{Clusters}
\title{Getter for clustering}
\usage{
Clusters(x, cluster, annotations = FALSE)
}
\arguments{
\item{x}{MOBSTER dataset.}

\item{cluster}{Use one of MOBSTER (M), sciClone (S), pyClone (P), Binomial (B).}

\item{annotations}{TRUE or FALSE will return also the latent variables if possible.}
}
\description{
Getter for clustering
}
