% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_getters.R
\name{Data_table}
\alias{Data_table}
\title{Tabular VAF, DP and NV values.}
\usage{
Data_table(x, ids = keys(x), samples = x$samples)
}
\arguments{
\item{x}{A \code{mbst_data} object}

\item{ids}{The IDs of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}
}
\value{
A spread tibble of the required entries.
}
\description{
This is just a wrapper to a combined call of function \code{VAF_table}, \code{DP_table} and 
\code{NV_table}. The resulting outputs are bound by column; ususal subset options are available.
}
