% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_fit_1D.R
\name{dbpmm.2steps.fit}
\alias{dbpmm.2steps.fit}
\title{Title}
\usage{
dbpmm.2steps.fit(X, output.folder = ".", fit.tail = TRUE,
  second.fit = c("DP", "vbdbmm"), do.plots = TRUE,
  annotation = "dbpmm.tumour", dbpmm.fit.params = list(K = 1, samples = 1,
  init = "peaks", tail = TRUE, epsilon = 1e-10, maxIter = 6000, is_verbose =
  FALSE, fit.type = "MM", parallel = F, cores.ratio = 0.8, file.dump = NA, seed
  = 12345, top = 10, annotation = NULL), DP.fit.params = list(alpha_0 = 1e-04,
  a1 = 1, b1 = 1, ngrid = 100, nburn = 5000, nsave = 10000, nskip = 3, ndisplay
  = 100, state = NULL, status = TRUE), vbdbmm.fit.params = list(K = 5, alpha_0
  = 1e-08, a_0 = 1, b_0 = 1, max_iter = NA, epsilon_conv = 1e-10, restarts = 10,
  parallel = TRUE, silent = TRUE))
}
\arguments{
\item{vbdbmm.fit.params}{}
}
\description{
Title
}
\examples{
will make some
}
