% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_fit_1D.R
\name{dbpmm.fit}
\alias{dbpmm.fit}
\title{dbpmm.fit}
\usage{
dbpmm.fit(X, K = 1:3, samples = 10, init = "peaks", tail = c(TRUE,
  FALSE), epsilon = 1e-10, maxIter = 2000, is_verbose = FALSE,
  fit.type = "MM", parallel = FALSE, cores.ratio = 0.8, file.dump = NA,
  seed = 12345, annotation = NULL, model.selection = "reICL")
}
\arguments{
\item{X}{Data, a vector of observations in (0,1).}

\item{K}{A vector with the number of Beta components to use. A single  Pareto component is used to model
tails, see also "tail". All values of K must be positive and strictly greater than 0.}

\item{samples}{Number of fits that should be attempted for each value of K. This value >1 makes senso only if
one uses randomized initial conditions, of course.}

\item{init}{Initial values for the paremeters. By using "ranodm" mean and variance for each Beta component
are randomply sampled in the interval (0,1). PARETO??? If this is a list with K-dimensional vectors named 'a' and 'b',
and scalar 'shape' and 'scale' are provided, these parameters will be used.}

\item{tail}{Set it FALSE to fit the data without the Pareto tail.}

\item{epsilon}{Tolerance for convergency estimation. For MLE fit this is compared to the differential of the NLL.
for MM fit to the largest differential among the mixing proportions (pi).}

\item{maxIter}{Maximum number of steps of the fit. If convergency is not achieved before these steps, the fit is interrupted.}

\item{is_verbose}{Verbose output. This also shows a sort of progress bar for the fitting.}

\item{fit.type}{A string that determines the type of fit. "MLE" is the Maximum Likelihood Estimate of the Beta paraneters,
while "MM" is Momemnt Matching. MLE is numerical, and thus slower.}

\item{parallel}{If TRUE, package parallel is used to run all fits in parallel.}

\item{cores.ratio}{Ratio (0,1) of the total number of cores that should be used if parallel = TRUE}

\item{file.dump}{If this is not NA then it should be a string. The method will save all fits in an RData named accordingly
and will plot all the fits in a PDF file.}

\item{seed}{Seed for the random numbers generator}

\item{annotation}{Subtitle annotation, if plotting.}

\item{model.selection}{Criterion to pick the best model -- one of ICL, BIC, AIC, NLL or entropy. We advise to use only ICL and BIC.}

\item{top}{Number of top fits to return, ranked by ICL.}
}
\value{
An object of class "dbpmm" which has methods for print and visualization.
}
\description{
dbpmm.fit
}
\examples{
will make some
}
