% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{ddbpmm}
\alias{ddbpmm}
\title{Density function for a Dirichlet Beta-Pareto Mixture Model}
\usage{
ddbpmm(x, data = NULL, components = 1:x$K, a = NULL, b = NULL,
  pi = NULL, scale = NULL, shape = NULL, log = TRUE)
}
\arguments{
\item{x}{an object of class dbpmm}

\item{data}{data to compute the (log)-likelihood, if this is null the data stored inside the object is used.}

\item{components}{a vector specifying which components should be used to compute the density (this is used in the implementation of the
fitting algorithms for this mixture -- most users would not use this option).}

\item{log}{log-likelihood or just likelihood.}

\item{init}{if TRUE, it uses the initial values for model fit, otherwise final.}
}
\value{
The density for the data. Notice, this is NOT the negative LL returned during fit (which is this value with different sign).
}
\description{
Density function for a Dirichlet Beta-Pareto Mixture Model
}
\examples{
Will make some..
}
