% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnds.R
\name{dnds}
\alias{dnds}
\title{Run a dN/dS analysis on Mobster clusters}
\usage{
dnds(fit, gene_list = NULL, refdb = "hg19")
}
\arguments{
\item{fit}{MOBSTER fit object}

\item{gene_list}{An optional vector of gene names to infer dN/dS values, 
default is to use the whole exome}

\item{refdb}{The genome referene to use, default is to use hg19. Other references are available from 
https://github.com/im3sanger/dndscv_data}
}
\value{
the fit object is returned with additional dnds list to the best fit model.
}
\description{
This function takes a MOBSTER fit and runs dndscv (https://github.com/im3sanger/dndscv) to calculate dN/dS 
values per cluster. It computes global dN/dS and per gene dN/dS values and makes a plot. dN/dS values are computed with
the best fitting MOBSTER model.
}
