% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_getters.R
\name{DP_table}
\alias{DP_table}
\title{Tabular DP values.}
\usage{
DP_table(x, ids = keys(x), samples = x$samples, suffix = ".DP")
}
\arguments{
\item{x}{A \code{mbst_data} object}

\item{ids}{The IDs of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}

\item{suffix}{Every column will be called as \code{sample_name.DP} when \code{suffix = '.DP'}}
}
\value{
A spread tibble of the required entries.
}
\description{
Similarly to \code{DP}, this function however returns a spread-like table 
with one column per VAF value. As \code{DP} the entries can be subset as
required, and the columns named with a particular format.
}
