% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evodynamics.R
\name{estimate_prior}
\alias{estimate_prior}
\title{Estimate the parameters of the mutation rate prior from a mobster fit.}
\usage{
estimate_prior(fit)
}
\arguments{
\item{fit}{HMobster fit}
}
\value{
a list containing the parameters of the prior distribution
}
\description{
It is assumed a gamma distribution as prior for the mutation rate. The parameters alpha and beta are estimated from
the data. We compute an estimator of mu for each karyotype, i.e. \mu_i= M_i/((1/f_{min;i}-1/f_{max;i})\ell_i) and take the mean and
variance accross different karyotypes. Using the MM formula for the gamma distribution we get an expression for
alpha and beta parameters.
}
\examples{
data('fit_example_mobsterh', package = 'mobster')
estimate_prior(fit_example_mobsterh)
}
