% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Animation.R
\name{fit_animation}
\alias{fit_animation}
\title{Render the animation of a MOBSTER fit}
\usage{
fit_animation(x, density = 0.05, lib = "plotly")
}
\arguments{
\item{x}{An object of class \code{"dbpmm"} computed with 
\code{trace = TRUE} in \code{mobster_fit}.}

\item{density}{The density of steps in the trace that should be used
to render the animation. For instance, 5% would mean that if there are
n points in the trace, one every n*0.05 is used to animate the fit.}

\item{lib}{Library used for renderining; this can be either 
\code{'plotly'} or \code{'gganimate'}.}
}
\value{
The animation object, according to the chosen  \code{lib}.
}
\description{
Render the animation of a MOBSTER fit
}
\examples{
data(fit_example)

model = mobster_fit(
fit_example$best$data, 
parallel =  FALSE, 
samples = 3,
init = 'random',
trace = TRUE,
K = 2,
tail = T)

fit_animation(model$best, lib = 'plotly')
}
