% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_example}
\alias{fit_example}
\title{Example fit}
\format{Data from a multivariate Binomial mixture}
\usage{
data(fit_example)
}
\description{
Data from a multivariate dimensional Binomial mixture
with unknown ground truth.
}
\examples{
data(fit_example)
head(fit_example)
}
\keyword{datasets}
