% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_example_mobsterh}
\alias{fit_example_mobsterh}
\title{Example MOBSTERh fit.}
\format{
Output from `mobsterh_fit`.
}
\usage{
data(fit_example_mobsterh)
}
\description{
Data from an example MOBSTERh fit; this object is the
result of running `mobsterh_fit` function on input data.
The object format is the same of `fit_example`.
}
\examples{
data(fit_example_mobsterh)
print(fit_example_mobsterh$best)
plot(fit_example_mobsterh$best)
}
\keyword{datasets}
