% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{format_input}
\alias{format_input}
\title{Prepare multivariate inputs.}
\usage{
format_input(data, CN, samples, VAF.range = c(0.05, 0.6), subsample = NULL)
}
\arguments{
\item{data}{input SNVs detected across all samples}

\item{CN}{Copy Number segments}

\item{samples}{samples ID}

\item{VAF.range}{VAF range for mutations to analyze}

\item{subsample}{VAF range for mutations to analyze}
}
\value{
data for input in sciClone
}
\details{
Requirement for a multivariate analysis, a data frame with this format

> head(data)
id  NV.B1  DP.B1       VAF.B1  NV.B2  DP.B2      VAF.B2       chr    pos
241    62    123    0.5040650     54    112   0.4821429    pseudo      1
253    46    100    0.4600000     57    108   0.5277778    pseudo      2
264    65    119    0.5462185     63    120   0.5250000    pseudo      3
275    48    112    0.4285714     68    129   0.5271318    pseudo      4
286    45    106    0.4245283     66    117   0.5641026    pseudo      5
297    59    116    0.5086207     69    136   0.5073529    pseudo      6
In the above dataset we have 2 samples, named "B1" and "B2". Input columns
"id", "chr" and "pos" are optional and are auto-generated in case they are
missing from the input dataframe.
}
