\name{get.hob}
\alias{get.hob}
\title{Get the historical book.}
\usage{get.hob(n)}
\description{Get the historical book.
Each quote/order is recorded in a historical book for analysis.
The historical book records 10 levels of bid/ask prices, qty,
the total traded volume and total number of trades}
\arguments{\item{n}{the number of orders to retrieve}}
