\name{get.ob}
\alias{get.ob}
\title{Get the orderbook.}
\description{Get the orderbook.
Gets a snapshot of the current orderbook. The order book has 10 ask and 10 bid levels.
By adding limit orders/quotes and market orders the dynamics of an exhange can be modelled.
By using a file of real exchange data trading strategies can be plugged in to model market microstructure.}
