% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LU4_lung_sample}
\alias{LU4_lung_sample}
\title{MOBSTER fit for the LU4 lung sample.}
\format{
Output from `mobster_fit`.
}
\usage{
data(LU4_lung_sample)
}
\description{
MOBSTER fit for the LU4 lung sample; this object is the
result of running `mobster_fit` function on the input data described in the 
main MOBSTER paper. The data consists of diploid mutations for the 
sample available at the Comprehensive Omics Archive of Lung Adenocarcinoma
(http://genome.kaist.ac.kr/).
}
\examples{
data(LU4_lung_sample, package = 'mobster')
print(LU4_lung_sample$best)
plot(LU4_lung_sample$best)
}
\keyword{datasets}
