% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUFF76_lung_sample}
\alias{LUFF76_lung_sample}
\title{MOBSTER fit for the PD4120a breast sample.}
\format{
Data from a MOBSTER fit.
}
\usage{
data(LUFF76_lung_sample)
}
\description{
MOBSTER fit for the LUFF76 lung sample; this object is the
result of running `mobster_fit` function on the input data described in the 
main MOBSTER paper.
}
\examples{
data(LUFF76_lung_sample, package = 'mobster')
print(LUFF76_lung_sample$best)
plot(LUFF76_lung_sample$best)
}
\keyword{datasets}
