% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{MOBSTER_animate}
\alias{MOBSTER_animate}
\title{Title}
\usage{
MOBSTER_animate(x, density = 0.05, transition_length = 1,
  state_length = 1, lib = "plotly", palette = "Set1",
  tail.color = "gray")
}
\arguments{
\item{x}{}

\item{density}{}

\item{palette}{}

\item{tail.color}{}
}
\description{
Title
}
