% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_bootstrap.R
\name{mobster_bootstrap}
\alias{mobster_bootstrap}
\title{Bootstrap a MOBSTER fit.}
\usage{
mobster_bootstrap(x, n.resamples = 100, bootstrap = "nonparametric",
  cores.ratio = 0.8, cache = NULL, save_data = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"dbpmm"}.}

\item{n.resamples}{Number of boostrap resamples.}

\item{bootstrap}{Type of boostrap: \code{"parametric"} or \code{"nonparametric"}}

\item{cores.ratio}{Ratio of cores to use for the parallel; see \code{?easypar}.}

\item{cache}{Cache for the computation; see \code{?easypar}.}

\item{...}{fit parameters for \code{mobster_fit}}
}
\value{
Data from the fits, resamples and a plottable figure.
}
\description{
Parametric and non-parametric implementation of
bootstrap estimates for MOBSTER fits. This computation is parallel
and uses  \code{?easypar}.
}
\examples{
data('fit_example', package = 'mobster')

# Just 5 resamples of a nonparametric bootstrap run
mb = mobster_bootstrap(fit_example$best, n.resamples = 5)

# The resample data is wrapped as lists of lists
lapply(mb$resamples, function(x) x[[1]] \%>\% as_tibble)

# The best fits are returned
mb$fits
}
