% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{mobster_bootstrap_cocluster}
\alias{mobster_bootstrap_cocluster}
\title{Co-clustering probability from non-parametric bootstrap}
\usage{
mobster_bootstrap_cocluster(resamples, fits, background.color = "white")
}
\arguments{
\item{resamples}{Data resampled}

\item{fits}{Fits (one per resample)}
}
\value{
Nothing
}
\description{
Compute co-clustering probability from non-parametric bootstrap resamples

This function can be applied only to nonparametric bootstrap resamples.
}
