% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{mobster_bootstrap_params}
\alias{mobster_bootstrap_params}
\title{Analyze fits by bootstrap.}
\usage{
mobster_bootstrap_params(resamples, fits)
}
\arguments{
\item{resamples}{Data resampled}

\item{fits}{Fits (one per resample)}
}
\value{
A figure
}
\description{
Computes model's stability (percentage of model
fits across resamples) and parameters' boxplots. 

This function can be applied also to both parametric and nonparametric bootstrap.
}
