% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset.R
\name{mobster_dataset}
\alias{mobster_dataset}
\title{Title}
\usage{
mobster_dataset(data, samples, description = "My own dataset", segments,
  purity, N.min = 500, na.rm = TRUE)
}
\arguments{
\item{data}{}

\item{samples}{}

\item{description}{}

\item{segments}{}

\item{purity}{}

\item{N.min}{}

\item{na.rm}{}
}
\description{
Title
}
