% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset.R
\name{mobster_fit_multivariate}
\alias{mobster_fit_multivariate}
\title{Title}
\usage{
mobster_fit_multivariate(x, samples = x$samples, ...)
}
\arguments{
\item{x}{}

\item{samples}{}

\item{...}{}
}
\description{
Title
}
