% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset.R
\name{MOBSTER_fit_pyClone}
\alias{MOBSTER_fit_pyClone}
\title{Title}
\usage{
MOBSTER_fit_pyClone(dataset, iterations = 10000, burnin = 1000,
  seed = 100)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Title
}
