% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset.R
\name{mobster_fit_sciClone}
\alias{mobster_fit_sciClone}
\title{Title}
\usage{
mobster_fit_sciClone(x, minimumDepth = 0, maximumClusters = 2 *
  length(x$samples), ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Title
}
