% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_filters.R
\name{mobster_flt_vafrange}
\alias{mobster_flt_vafrange}
\title{Title}
\usage{
mobster_flt_vafrange(x, min.vaf, max.vaf)
}
\arguments{
\item{x}{}

\item{min.vaf}{}

\item{max.vaf}{}
}
\description{
Title
}
