% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plots_multivariate.R
\name{MOBSTER_mplot_fits_MOBSTER_sciClone}
\alias{MOBSTER_mplot_fits_MOBSTER_sciClone}
\title{Title}
\usage{
MOBSTER_mplot_fits_MOBSTER_sciClone(data, samples,
  cluster1 = "sciClone.cluster", cluster2 = "MOBSTER.sciClone.cluster",
  palette.cluster1 = "Set1", palette.cluster2 = "Spectral", cex = 1,
  VAF.range = c(0.05, 0.6))
}
\arguments{
\item{VAF.range}{}
}
\description{
Title
}
