% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plots_multivariate.R
\name{MOBSTER_mplot_mapping2MOBSTER_clusters}
\alias{MOBSTER_mplot_mapping2MOBSTER_clusters}
\title{Title}
\usage{
MOBSTER_mplot_mapping2MOBSTER_clusters(data, MOBSTER.fits,
  cluster = "sciClone.cluster", samples, palette = "Set2", cex = 1,
  tail.color = "gray", title = paste0("Mapping of ", cluster,
  " clusters to MOBSTER ones"))
}
\arguments{
\item{title}{}
}
\description{
Title
}
