% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plots_multivariate.R
\name{MOBSTER_mplot_peaks}
\alias{MOBSTER_mplot_peaks}
\title{Title}
\usage{
MOBSTER_mplot_peaks(data, cluster, VAF.samples, title = "Clone peaks",
  cex = 1, cutoff = 0.05, palette = "Set1", free = FALSE)
}
\arguments{
\item{free}{}
}
\description{
Title
}
