% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_2D_cluster_density.R
\name{mobster_plt_2D_cluster_density}
\alias{mobster_plt_2D_cluster_density}
\title{Title}
\usage{
mobster_plt_2D_cluster_density(x, cov = 100, palette = "Set1", ...)
}
\arguments{
\item{x}{}

\item{palette}{}

\item{...}{}
}
\description{
Title
}
