% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_2DVAF_matrix.R
\name{mobster_plt_2DVAF_matrix}
\alias{mobster_plt_2DVAF_matrix}
\title{Title}
\usage{
mobster_plt_2DVAF_matrix(lower.x, lower.cluster = NULL,
  lower.cluster.label = NULL, lower.exclude = NULL, upper.x = NULL,
  upper.cluster = NULL, upper.cluster.label = NULL,
  upper.exclude = NULL, MOBSTER = NULL, samples = lower.x$samples,
  lower.description = lower.x$description,
  upper.description = ifelse(is.null(upper.x), "", upper.x$description),
  palette = list(MOBSTER = "Set1", lower = "Set2", upper = "Spectral"),
  bg = list(MOBSTER = "gainsboro", lower = "darkslategray4", upper =
  "ivory3"), ...)
}
\arguments{
\item{lower.cluster}{}

\item{lower.cluster.label}{}

\item{upper.cluster}{}

\item{upper.cluster.label}{}

\item{MOBSTER}{}

\item{palette}{}

\item{...}{}

\item{x}{}

\item{title}{}
}
\description{
Title
}
