% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\name{mobster_plt_bootstrap_params}
\alias{mobster_plt_bootstrap_params}
\title{Plots bootstrap fits of model and parameters.}
\usage{
mobster_plt_bootstrap_params(resamples, fits, tail.color = "darkgray",
  cex = 1, palette = "Set1")
}
\arguments{
\item{resamples}{Data resampled}

\item{fits}{Fits (one per resample)}

\item{tail.color}{}
}
\value{
A figure
}
\description{
Plots model's stability (percentage of model
fits across resamples) and bootstrap estimates of the parameters' boxplots. 

This function can be applied also to both parametric and nonparametric bootstrap.
}
