% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_clustering_assignments_summary.R
\name{mobster_plt_clustering_assignment_summary}
\alias{mobster_plt_clustering_assignment_summary}
\title{Title}
\usage{
mobster_plt_clustering_assignment_summary(fit_with_MOBSTER,
  fit_without_MOBSTER, CCF.cutoff = 0.7,
  title = fit_with_MOBSTER$description, cex_legend = 1)
}
\arguments{
\item{fit_with_MOBSTER}{}

\item{fit_without_MOBSTER}{}

\item{CCF.cutoff}{}

\item{title}{}

\item{cex_legend}{}
}
\description{
Title
}
