% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_CNsegments.R
\name{mobster_plt_CNsegments}
\alias{mobster_plt_CNsegments}
\title{Title}
\usage{
mobster_plt_CNsegments(x, samples = x$samples, ...)
}
\arguments{
\item{x}{}

\item{samples}{}

\item{...}{}
}
\description{
Title
}
