% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_filters.R
\name{mobster_plt_filters}
\alias{mobster_plt_filters}
\title{Plots mutations that would be fitered with MOBSTER function.}
\usage{
mobster_plt_filters(x, VAF_min, NV_min, min.DP, max.DP, min.VAF, max.VAF,
  x.lim = NA, y.lim = NA)
}
\arguments{
\item{x}{A MOBSTER dataset.}

\item{VAF_min}{VAF values below this parameter are considered as 0 (undetected);
see function \code{\link{mobster_flt_minvaf}}. This filter applies to each
sample independently.}

\item{NV_min}{NV values below this parameter are considered as 0 (undetected);
see function \code{\link{mobster_flt_minnv}}. This filter applies to each
sample independently.}

\item{min.DP}{The depth value for each mutation must be higher than \code{min.DP},
whenever it is greater than 0 in a sample; see function \code{\link{mobster_flt_dprange}}.}

\item{max.DP}{The depth value for each mutation must be lower than \code{max.DP};
see function \code{\link{mobster_flt_dprange}}.}

\item{min.VAF}{The VAF value for each mutation must be higher than \code{min.VAF},
whenever it is greater than 0 in a sample; see function \code{\link{mobster_flt_vafrange}}.}

\item{max.VAF}{The depth value for each mutation must be lower than \code{min.VAF};
see function \code{\link{mobster_flt_vafrange}}.}

\item{x.lim}{X-axis limits for the plot (e.g., \code{c(0,1)}); NA is no limits.}

\item{y.lim}{Y-axis limits for the plot (e.g., \code{c(0,1)}); NA is no limits.}
}
\value{
A set of figures (per filter, and per sample) with all annotated mutations
coloured according to their status (if they would be removed by a filter, or not).
}
\description{
MOBSTER has functions to subset the mutations
that one has annotated in the data. With this function one
can generate plots that show which mutations would be filtered
for a certain set of parameters, this allows to one to tune
the filters available in the too.
}
