% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_model_selection.R
\name{mobster_plt_model_selection}
\alias{mobster_plt_model_selection}
\title{Title}
\usage{
mobster_plt_model_selection(res, fig.lab = "",
  title = "MOBSTER best fit", TOP = 5, cex = 1,
  boxplot.ICL.range = NULL, ...)
}
\arguments{
\item{boxplot.ICL.range}{}
}
\description{
Title
}
