% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset.R
\name{mobster_purity}
\alias{mobster_purity}
\title{Title}
\usage{
mobster_purity(x, peak.range = c(0.2, 0.5), m.peak = 10)
}
\arguments{
\item{x}{}

\item{peak.range}{}

\item{m.peak}{}
}
\description{
Title
}
