% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_wrp_neutralitytestr.R
\name{mobster_wrp_neutralitytest}
\alias{mobster_wrp_neutralitytest}
\title{Title}
\usage{
mobster_wrp_neutralitytest(x, samples = x$samples, ql = 0.05,
  qr = 0.95)
}
\arguments{
\item{samples}{}
}
\description{
Title
}
