\name{model.poisson}
\alias{model.poisson}
\title{Model the book as Poisson events.}
\usage{model.poisson(num=100)}
\description{Model the book as Poisson events.
This demonstrates constructing a book from limit orders and trades
generated by a poisson process.}
\arguments{\item{num}{The number of iterations to model}}
