% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.selection.R
\name{model_selection}
\alias{model_selection}
\title{Title}
\usage{
model_selection(results, scores.suitable = c("ICL", "BIC", "AIC", "NLL",
  "reICL"))
}
\arguments{
\item{scores.suitable}{}
}
\description{
Title
}
