% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evodynamics.R
\name{mu_posterior}
\alias{mu_posterior}
\title{Estimate mutation rate posterior from MOBSTER fit}
\usage{
mu_posterior(
  fit,
  genome_length,
  quantiles = c(0.02, 0.98),
  prior = list(alpha = 1e-04, beta = 1e-04)
)
}
\description{
@description The parameters defining the posterior distribution of the mutation rate are computed, together with the plot
 of the distribution. The accumulation of subclonal mutations on the genome is described by a Poisson
 process with intensity \Lambda=\mu\sum_{i \in mathrm{karyotypes}}(1/f_{min;i}-1/f_{max;i})\ell_i,
 where \mu is the effective mutation rate, f_{min;i},f_{max;i} are respectively the minimum and maximum frequency
 of the mutations assigned to the tail and \ell_i is the length of the genomic region for a given karyotype.
 Assuming a gamma distribution with parameters alpha,beta as prior, the posterior distribution is again a
 gamma distribution with parameters \alpha^{\prime} = \alpha + \sum_{i in \mathrm{karyotypes}}M_i,
 \beta = \beta + \sum(1/f_{min;i}-1/f_{max;i})\ell_i, where M_i are the number of subclonal mutations for a given karyotype.
 @param fit Fit by MOBSTERh
 @param prior Object containing the alpha and beta parameters of the gamma prior
 @param genome_length
 @return a list containing the parameters of the posterior distribution and a density plot
 @examples
 data('fit_example_mobsterh', package = 'mobster')
 prior=list(alpha=10^-4,beta=10^-4)
 mu_posterior(fit_example_mobsterh,prior=prior)
 @export
}
