% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evodynamics.R
\name{mutationrate}
\alias{mutationrate}
\title{Estimate mutation rate from MOBSTER fit}
\usage{
mutationrate(fit, lq = 0.025, uq = 0.975)
}
\arguments{
\item{fit}{Mobster fit}
}
\value{
Mutation rate per tumour doubling
}
\description{
Mutation rate is calculated based on population genetics model, 
for details see Williams et al. In summary the mutation rate scaled
by the probability of lineage survival \eqn{\beta}, \eqn{\mu/\beta} is given by:
\deqn{\mu/\beta = M / (1/fmin - 1/fmax)}
Where \eqn{fmin} is the minimum VAF and \eqn{fmax} is the maximum, and 
\eqn{M} is the number of mutations between \eqn{fmin} and \eqn{fmax}.
}
\examples{
mutationrate(mobsterfit)
}
