% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_getters.R
\name{NV}
\alias{NV}
\title{Extract NV values.}
\usage{
NV(x, ids = keys(x), samples = x$samples)
}
\arguments{
\item{x}{A \code{mbst_data} object}

\item{ids}{The IDs of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}
}
\value{
A tibble of the required entries
}
\description{
Extract from the internal representation of an object all the entries
that refer to the NV values. These can be subset by sample and mutation
ID; by default all entries are returbed. The output is a tibble; no other
transformations are executed.
}
