% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PD4120a_breast_sample}
\alias{PD4120a_breast_sample}
\title{MOBSTER fit for the PD4120a breast sample.}
\format{Data from an example MOBSTER fit.}
\usage{
data(PD4120a_breast_sample)
}
\description{
MOBSTER fit for the PD4120a breast sample; this object is the
result of running `mobster_fit` function on the input data described in the 
main MOBSTER paper.
}
\examples{
data(PD4120a_breast_sample, package = 'mobster')
print(PD4120a_breast_sample$best)
plot(PD4120a_breast_sample$best)
}
\keyword{datasets}
