% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_plot.R
\name{plot.dbpmm}
\alias{plot.dbpmm}
\title{Plot a MOBSTER fit.}
\usage{
plot.dbpmm(x, alpha = 0.8, cex = 1, colors = NA, ...)
}
\arguments{
\item{x}{An object of class \code{"dbpmm"}.}

\item{alpha}{Alpha value for the colors of the histogram}

\item{cex}{Cex of the plot}

\item{colors}{If provided, these colours will be used for each cluster.
If not all clusters have available colours, errors are thrown.}

\item{...}{}
}
\value{
A ggplot object for the plot.
}
\description{
Plot a MOBSTER fit.
}
\examples{
something..
}
