% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobsterm_methods_plot.R
\name{plot.dbpmm_m}
\alias{plot.dbpmm_m}
\title{Scatterplots of a MOBSTERm fit.}
\usage{
plot.dbpmm_m(x)
}
\arguments{
\item{x}{A MOBSTERm fit.}
}
\value{
A list of ggplot objects for the scatter plots.
}
\description{
Scatterplots of a MOBSTERm fit.
}
\examples{
TBD
}
