% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_plot.R
\name{plot.mbst_data}
\alias{plot.mbst_data}
\title{Plot a dataset.}
\usage{
plot.mbst_data(x, title = x$description, scales = "fixed")
}
\arguments{
\item{x}{}

\item{title}{Title of the plot, default is taken from \code{x}.}

\item{scales}{Scales for ggplot facets, default is \code{'fixed'} but \code{'free'} can also help.}
}
\value{
A ggpubr figure object.
}
\description{
Plot a dataset.
}
\details{
It shows the VAF distributions of the data, with a special panel for
low-frequency VAF values, and panels for coverage (DP). Number of reads with the
mutant allele (NV) are not shown as they are as the VAF. There is an option to
plot the VAF value by kariotype.
}
