% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plots_multivariate.R
\name{plot_2DVAF}
\alias{plot_2DVAF}
\title{Title}
\usage{
plot_2DVAF(data, x, y, cluster = NULL, marginal = FALSE, cex = 1,
  VAF.range = c(0.05, 0.6), marginal.remove.zeroes = T, palette = "Set1")
}
\arguments{
\item{marginal}{}
}
\description{
Title
}
