% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boostrap_Beta.R
\name{plot_bootstrap_Beta}
\alias{plot_bootstrap_Beta}
\title{Plot the boostrapped tail parameters.}
\usage{
plot_bootstrap_Beta(x, bootstrap_results, bootstrap_statistics,
  colors = c(Tail = "gainsboro"))
}
\arguments{
\item{x}{A MOBSTER fit.}

\item{bootstrap_results}{Results of a call to \code{mobster_boostrap}.}

\item{colors}{If provided, these colours will be used for each cluster.
If a subset of colours is provided, palette Set1 from \code{RColorBrewer} is used.
By default the tail colour is provided as 'gainsboro'.}

\item{bootstrap_statisticsResults}{of a call to \code{bootstrapped_statistics}.}
}
\value{
A barplot of the bootstrapped tail parameters.
}
\description{
From the results of a call to \code{mobster_boostrap}, and
the results of a call to \code{bootstrapped_statistics}, a boxplot with
the tail parameters is produced.
}
