% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boostrap_coclustering.R
\name{plot_bootstrap_coclustering}
\alias{plot_bootstrap_coclustering}
\title{Plot the boostrapped co-clustering probability.}
\usage{
plot_bootstrap_coclustering(
  x,
  bootstrap_results,
  bootstrap_statistics,
  colors = c(Tail = "gainsboro")
)
}
\arguments{
\item{x}{A MOBSTER fit.}

\item{bootstrap_results}{Results of a call to \code{mobster_boostrap}.}

\item{colors}{If provided, these colours will be used for each cluster.
If a subset of colours is provided, palette Set1 from \code{RColorBrewer} is used.
By default the tail colour is provided as 'gainsboro'.}

\item{bootstrap_statisticsResults}{of a call to \code{bootstrapped_statistics}.}
}
\value{
A \code{cowplot} figure of the bootstrapped co-clustering probability.
}
\description{
From the results of a call to \code{mobster_boostrap}, and
the results of a call to \code{bootstrapped_statistics}, a figure (heatmap-style)
for the boostrapped co-clustering probability is produced assembling plots 
with \code{cowplot::plot_grid}.
}
