% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boostrap_model_frequency.R
\name{plot_bootstrap_model_frequency}
\alias{plot_bootstrap_model_frequency}
\title{Plot the boostrapped model frequency}
\usage{
plot_bootstrap_model_frequency(bootstrap_results, bootstrap_statistics)
}
\arguments{
\item{bootstrap_results}{Results of a call to \code{mobster_boostrap}.}

\item{bootstrap_statisticsResults}{of a call to \code{bootstrapped_statistics}.}
}
\value{
A barplot of the boostrapped model frequency
}
\description{
From the results of a call to \code{mobster_boostrap}, and
the results of a call to \code{bootstrapped_statistics}, a barplot with
the model frequency is produced
}
