% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{plot_data}
\alias{plot_data}
\title{Title}
\usage{
plot_data(x, VAF.columns, DP.columns, NV.columns,
  title = "Raw data: VAF (full and low-freq.), DP, NV")
}
\arguments{
\item{title}{}
}
\description{
Title
}
