% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entropy.R
\name{plot_entropy}
\alias{plot_entropy}
\title{Plot the entropy of a MOBSTER mixture.}
\usage{
plot_entropy(x)
}
\arguments{
\item{x}{An object of class \code{"dbpmm"}.}
}
\value{
A ggplot object for the plot.
}
\description{
Returns a plot of the entropy and the reduced entropy
for this mixture, binning the domain with bins of size 1e-3. The
full entropy is coloured with a gradient, and the reduced is dashed.
}
\examples{
data(fit_example)
plot_entropy(fit_example$best)
}
