% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_gofit}
\alias{plot_gofit}
\title{Plot the goodness of fit.}
\usage{
plot_gofit(x, TOP = 5, cex = 1)
}
\arguments{
\item{x}{A list of fits computed via \code{mobster_fit}.}

\item{TOP}{The top set of fits to use.}

\item{cex}{Cex of the plot.}
}
\value{
A plot of the  goodness of fit.
}
\description{
Plot the SSE (sum of squared error) as a proxy for the
goodness of fit. For the \code{TOP} solutions the SSE trace is shown.
}
\examples{
TODO
}
