% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{plot_grid_plain}
\alias{plot_grid_plain}
\title{Grid plot for multivariate analysis.}
\usage{
plot_grid_plain(data, samples, VAF.range = c(0.05, 0.6), cex = 1,
  title = "VAF pairwise plot")
}
\arguments{
\item{samples}{}
}
\description{
Grid plot for multivariate analysis.
}
