% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_latent_variables}
\alias{plot_latent_variables}
\title{Plot the latent variables of the mixture.}
\usage{
plot_latent_variables(x, cex = 1)
}
\arguments{
\item{x}{An object of class 'dbpmm'.}

\item{cex}{Cex of the plot.}
}
\value{
A plot of the latent variables of the mixture.
}
\description{
Plot the latent variables of the mixture.
}
\examples{
TODO
}
