% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobsterm_methods_plot_marginals.R
\name{plot_marginals.dbpmm_m}
\alias{plot_marginals.dbpmm_m}
\title{Marginal plots of a MOBSTERm fit.}
\usage{
plot_marginals.dbpmm_m(x)
}
\arguments{
\item{x}{A MOBSTERm fit.}
}
\value{
A list of ggplot objects for the marginal plots.
}
\description{
Marginal plots of a MOBSTERm fit.
}
\examples{
TBD
}
