% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_mixing_proportions}
\alias{plot_mixing_proportions}
\title{Plot the mixing proportions of the mixture.}
\usage{
plot_mixing_proportions(x, cex = 1)
}
\arguments{
\item{x}{An object of class 'dbpmm'.}

\item{cex}{Cex of the plot.}

\item{colors}{Optional vector of colors, default (\code{NA}) are \code{ggplot} colors.}
}
\value{
A plot of the mixing proportions of the mixture.
}
\description{
Plot the mixing proportions of the mixture.
}
\examples{
TODO
}
