% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_model_selection}
\alias{plot_model_selection}
\title{Plot summary for model selection.}
\usage{
plot_model_selection(x, TOP = 5, ...)
}
\arguments{
\item{x}{A list of fits computed via \code{mobster_fit}.}

\item{TOP}{The top set of fits to use.}
}
\value{
A complex figure with all plots arranged.
}
\description{
This plot is usefull to understand the different
fits and their rank with respect to some scoring used to
select the best model. The plot shows alternative solutions
and their rank as well.
}
\examples{
TODO
}
