% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NLL.R
\name{plot_NLL}
\alias{plot_NLL}
\title{Plot the NLL trace.}
\usage{
plot_NLL(x)
}
\arguments{
\item{x}{A MOBSTER fit.}
}
\value{
A plot of the the NLL trace for the input fit.
}
\description{
It plots the Negative log-Likelihood (NLL) trace of a fit.
}
\examples{
data('fit_example', package = 'mobster')
plot_NLL(fit_example$best)
}
