% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_report_MOBSTER}
\alias{plot_report_MOBSTER}
\title{Title}
\usage{
plot_report_MOBSTER(res, fig.lab = "Figure 1", title = "MOBSTER top fit",
  palette = "Set1", TOP = 5, cex = 1, boxplot.ICL.range = NULL)
}
\arguments{
\item{boxplot.ICL.range}{}
}
\description{
Title
}
