% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobsterm_methods_plot.R
\name{plotm}
\alias{plotm}
\title{Plot a MOBSTERm fit.}
\usage{
plotm(x)
}
\arguments{
\item{x}{A MOBSTERm fit.}
}
\value{
A list of ggplot objects for the plot.
}
\description{
Plot a MOBSTERm fit.
}
