% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.methods.R
\name{print.dbpmm}
\alias{print.dbpmm}
\title{Print an object of class dbpmm.}
\usage{
\method{print}{dbpmm}(x, ...)
}
\arguments{
\item{x}{an object of class dbpmm..}

\item{...}{}
}
\value{
nothing.
}
\description{
Print an object of class dbpmm.
}
\examples{
something..
}
