% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobsterh_methods_print.R
\name{print.dbpmmh}
\alias{print.dbpmmh}
\title{Print a MOBSTERh object.}
\usage{
\method{print}{dbpmmh}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'dbpmmh'}.}

\item{...}{}
}
\value{
nothing.
}
\description{
Print a MOBSTERh object.
}
\examples{
data("fit_example_mobsterh", package = 'mobster')
}
