% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset.R
\name{print.mbst_data}
\alias{print.mbst_data}
\title{Title}
\usage{
\method{print}{mbst_data}(x, show.log = TRUE, cuts.DP = c(50, 100),
  cuts.NV = c(5, 15), cuts.VAF = c(0.05, 0.15), ...)
}
\arguments{
\item{x}{}

\item{show.log}{}

\item{...}{}
}
\description{
Title
}
