% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobster_deconv_methods_print.R
\name{print.mobster_deconv}
\alias{print.mobster_deconv}
\title{Print a MOBSTER[h] model-selection object.}
\usage{
\method{print}{mobster_deconv}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'mobster_deconv'}.}

\item{...}{}
}
\value{
nothing.
}
\description{
Print a MOBSTER[h] model-selection object.
}
\examples{
data("fit_example_mobsterh", package = 'mobster')
}
