% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dataset.R
\name{random_dataset}
\alias{random_dataset}
\title{Generate a random MOBSTER model and data.}
\usage{
random_dataset(N = 5000, K_betas = 2, pi_tail_bounds = c(0.2, 0.4),
  pi_min = 0.1, Betas_separation = 0.1, Beta_variance_scaling = 1000,
  Beta_bounds = c(0.1, 0.9), seed = NULL)
}
\arguments{
\item{N}{Number of samples to generate (mutations).}

\item{K_betas}{Number of Beta components (subclones).}

\item{pi_tail_bounds}{2D vector with min and max size of the tail's mutations (proportions).}

\item{pi_min}{Minimum mixing proportion for every component.}

\item{Betas_separation}{Minimum separation between the means of the Beta components.}

\item{Beta_variance_scaling}{The variance of the Beta is generated as U[0,1] and scaled by this value.
Values on the order of 1000 give low variance, 100 represents a dataset with quite some dispersion (
compared to a putative Binomial generative model).}

\item{Beta_bounds}{Range of values to sample the Beta means.}

\item{seed}{The seed to fix the process, default is 123.}
}
\value{
A list with the dataset in a tibble, a plot the model's parameters.
}
\description{
Generate a random MOBSTER model, its data and creates a plot for it.
}
\examples{
x = random_dataset()


print(x)
}
