% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dataset_mobsterh.R
\name{random_dataset_mobsterh}
\alias{random_dataset_mobsterh}
\title{Generate a random MOBSTER model and data.}
\usage{
random_dataset_mobsterh(
  N = 5000,
  karyotypes_p = c(`1:0` = 1, `2:0` = 1, `1:1` = 6, `2:1` = 2, `2:2` = 1),
  breakpoints_rate = 2,
  reference_genome = "GRCh38",
  sex = "m",
  purity = 1,
  pi_min = 0.1,
  Betas_separation = 0.1,
  Beta_variance_scaling = 1000,
  Beta_bounds = c(0.1, 0.9),
  shape_bounds = c(1, 1, 3),
  scale = 0.05,
  seed = NULL
)
}
\arguments{
\item{N}{Number of samples to generate (mutations).}

\item{karyotypes_p}{A named vector of multinomial coefficients
in the form \code{`M:m` = p} denoting the un-normalised probability
mass of generating a segment with \code{"M"} copies of the major allele,
and \code{"m"} copies of the minor allele.}

\item{breakpoints_rate}{Poisson rate $\lambda>0$ to sample the number of
breakpoints per chromosome. The sampling process ensures that}

\item{reference_genome}{Reference genome ("hg19 or hg38") to use as reference}

\item{sex}{sex of the individual}

\item{pi_min}{Minimum mixing proportion for every component.}

\item{Betas_separation}{Minimum separation between the means of the Beta components.}

\item{Beta_variance_scaling}{The variance of the Beta is generated as U[0,1] and scaled by this value.
Values on the order of 1000 give low variance, 100 represents a dataset with quite some dispersion (
compared to a putative Binomial generative model).}

\item{Beta_bounds}{Range of values to sample the Beta means.}

\item{shape_bounds}{Range of values to sample the tail shape, default [1, 3],}

\item{scale}{Tail scale, default 0.05.}

\item{seed}{The seed to fix the process, default is 123.}

\item{K_betas}{Number of Beta components (subclones).}

\item{pi_tail_bounds}{2D vector with min and max size of the tail's mutations (proportions).}
}
\value{
A list with the dataset in a tibble, the model parameters and a plot the data.
}
\description{
Generate a random MOBSTER model, its data and creates a plot for it.
}
\examples{
x = random_dataset()
print(x)
}
