% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{rdbpmm}
\alias{rdbpmm}
\title{Generate a random sample from a DBPMM model.}
\usage{
rdbpmm(x, a = NULL, b = NULL, pi = NULL, shape = NULL, scale = NULL,
  n = 1, tail.cutoff = 1)
}
\arguments{
\item{x}{an object with computed from MOBSTER}

\item{a}{if passed, `a` for the Beta components}

\item{b}{if passed, `a` for the Beta components}

\item{pi}{if passed, `pi` for the mixing proportions}

\item{shape}{if passed, `shape` of the tail}

\item{scale}{if passed, `scale` of the tail}

\item{n}{number of samples required}

\item{tail.cutoff}{reject Pareto Type I samples above this value (ensure that will be returned
the number of samples required anyway).}
}
\value{
n samples from the mixture
}
\description{
Generate a random sample from a DBPMM model.
}
\examples{
a = b = 50
names(a) = names(b) = "C1"

pi = c('Tail' = .6, 'C1' = .4)

v = rdbpmm(x = NULL, n = 1000, a = a, b = b, pi = pi, shape = 2, scale = 0.05)
hist(v, breaks = seq(0, 1, 0.01))
}
