% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_mobsterh.R
\name{sample_CNA_segments}
\alias{sample_CNA_segments}
\title{Sample an allele-specific CNA profile.}
\usage{
sample_CNA_segments(
  karyotypes_p = c(`1:0` = 1, `2:0` = 1, `1:1` = 6, `2:1` = 2, `2:2` = 1),
  breakpoints_rate = 2,
  reference_genome = "GRCh38",
  sex = "m"
)
}
\arguments{
\item{karyotypes_p}{A named vector of multinomial coefficients
in the form \code{`M:m` = p} denoting the un-normalised probability
mass of generating a segment with \code{"M"} copies of the major allele,
and \code{"m"} copies of the minor allele.}

\item{breakpoints_rate}{Poisson rate $\lambda>0$ to sample the number of
breakpoints per chromosome. The sampling process ensures that}

\item{sex}{}
}
\description{
Returns a chromosome-level segmentation of a genome
(hg19 or hg38) with associated allele-specific clonal Copy Number
Alterations. The karyotypes probabilities (multinomials) are given in
input, together with a Poisson rate parameter to generate breakpoints.

The function returns a tibble reporting information about the
sampled copy number segments.
}
