% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evodynamics.R
\name{selection}
\alias{selection}
\title{Calculate strength of selection of subclone}
\usage{
selection(time, time_end, subclonefrequency)
}
\arguments{
\item{time}{time subclone emerges (in tumour doublings)}

\item{time_end}{Time when tumour is sampled (in tumour doublings)}

\item{subclonefrequency}{Frequency of subclones}
}
\value{
s
}
\description{
Use properties of subclone fit to calculate selection intensity, selection 
is defined as the relative growth rates of host tumour cell 
populations (\eqn{\lambda h}) vs subclone (\eqn{\lambda s}):
\deqn{1+s=\lambda h / \lambda s}
}
\examples{
subcloneparameters(mobsterfit, mutationrate(mobsterfit), 1)
}
