% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evodynamics.R
\name{selection2clonenested}
\alias{selection2clonenested}
\title{Calculate strength of selection for 2 nested subclones}
\usage{
selection2clonenested(time1, time2, time_end, subclonefrequency1,
  subclonefrequency2)
}
\arguments{
\item{time1}{time subclone 1 emerges (in tumour doublings)}

\item{time2}{time subclone 2 emerges (in tumour doublings)}

\item{time_end}{Time when tumour is sampled (in tumour doublings)}

\item{subclonefrequency1}{Frequency of subclone 1}

\item{subclonefrequency2}{Frequency of subclone 2}
}
\value{
s
}
\description{
Use properties of subclone fit to calculate selection intensity, 
selection is defined as the relative growth rates of host tumour cell 
populations (\eqn{\lambda h}) vs subclone (\eqn{\lambda s}):
\deqn{1+s=\lambda h / \lambda s}
}
\examples{

}
