% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evodynamics.R
\name{subcloneparameters}
\alias{subcloneparameters}
\title{Extract relevent parameters from MOBSTER fit}
\usage{
subcloneparameters(fit, mu, subclonenumber = 1)
}
\arguments{
\item{fit}{Mobster fit}

\item{mu}{Mutation rate}

\item{subclonenumber}{ID of subclone}
}
\value{
tibble with all paramteres
}
\description{
Extract the number of mutations in the subclone, the frequency of the subclone
and calculate the time the subclone emerges.
}
\examples{
subcloneparameters(mobsterfit, mutationrate(mobsterfit), 1)
}
