% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.methods.R
\name{summary.dbpmm}
\alias{summary.dbpmm}
\title{Summaries for an object of class dbpmm is like a print.}
\usage{
\method{summary}{dbpmm}(x, ...)
}
\arguments{
\item{x}{the obj of class dbpmm}

\item{...}{}
}
\value{
see print
}
\description{
Summaries for an object of class dbpmm is like a print.
}
\examples{
something..
}
