% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobsterh_methods_print.R
\name{summary.dbpmmh}
\alias{summary.dbpmmh}
\title{Summary for an object of class \code{'dbpmmh'} is a print.}
\usage{
summary.dbpmmh(object, ...)
}
\arguments{
\item{object}{An obj of class \code{'dbpmmh'}.}

\item{...}{}
}
\value{
See \code{\link{print}}.
}
\description{
Summary for an object of class \code{'dbpmmh'} is a print.
}
\examples{
data(fit_example)
summary(fit_example$best)
}
