% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobster_deconv_methods_print.R
\name{summary.mobster_deconv}
\alias{summary.mobster_deconv}
\title{Summary for an object of class \code{'mobster_deconv'} is a print.}
\usage{
summary.mobster_deconv(object, ...)
}
\arguments{
\item{object}{An obj of class \code{'mobster_deconv'}.}

\item{...}{}
}
\value{
See \code{\link{print}}.
}
\description{
Summary for an object of class \code{'mobster_deconv'} is a print.
}
\examples{
data(fit_example)
summary(fit_example)
}
