% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_string.R
\name{to_string}
\alias{to_string}
\title{Return a tabular representation of a model.}
\usage{
to_string(x)
}
\arguments{
\item{x}{A MOBSTER fit.}
}
\value{

}
\description{
This functionr returns a table with all the parameters fits (one per column), the scores
of the model, and the SSE of the fit versus data.
}
\examples{
data('fit_example', package = 'mobster')

to_string(fit_example$best)
}
