\name{trade.mavg.sumpnl}
\alias{trade.mavg.sumpnl}
\title{sum the profit and loss...}
\usage{trade.mavg.sumpnl(x)}
\description{sum the profit and loss
perform some analysis on the moving avg strategy. It uses the trader id
for the different scenarios traded (cs='close sell',cb='close buy',os='open sell',ob='open buy')
and calculates where the trade was entered and exited.}
\arguments{\item{x}{the executions we made}}
