\name{trade.mavg}
\alias{trade.mavg}
\title{simple mavg trading for comparison (ie to show how terribly it performs!)...}
\usage{trade.mavg(num=100)}
\description{simple mavg trading for comparison (ie to show how terribly it performs!)
generate quote events according to poisson, and then use moving avg to enter/exit positions}
\arguments{\item{num}{The number of iterations to model}}
