\name{trade.mm.analyse}
\alias{trade.mm.analyse}
\title{analyse the market making strategy trades.}
\usage{trade.mm.analyse(numobs)}
\description{analyse the market making strategy trades.
this function gets all the executions and extracts the deals done by
the market making function. The pnl analysis calculates the spread pnl (how much is earned by providing liquidity)
and the costs of inventory holding.}
\arguments{\item{numobs}{The number of observations to analyse}}
