% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_getters.R
\name{VAF_table}
\alias{VAF_table}
\title{Tabular VAF values.}
\usage{
VAF_table(x, ids = keys(x), samples = x$samples, suffix = ".VAF")
}
\arguments{
\item{x}{A \code{mbst_data} object}

\item{ids}{The IDs of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}

\item{suffix}{Every column will be called as \code{sample_name.VAF} when \code{suffix = '.VAF'}}
}
\value{
A spread tibble of the required entries.
}
\description{
Similarly to \code{VAF}, this function however returns a spread-like table 
with one column per VAF value. As \code{VAF} the entries can be subset as
required, and the columns named with a particular format.
}
