% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_dataset_getters.R
\name{Annotations}
\alias{Annotations}
\title{Title}
\usage{
Annotations(x, ids = keys(x), variable = unique(x$annotations$variable),
  samples = x$samples)
}
\arguments{
\item{x}{}

\item{ids}{}

\item{variable}{}

\item{samples}{}
}
\description{
Title
}
