% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_statistics.R
\name{bootstrapped_statistics}
\alias{bootstrapped_statistics}
\title{Compute boostrap statistics from a bootstrap run.}
\usage{
bootstrapped_statistics(
  x,
  bootstrap_results,
  bootstrap = "nonparametric",
  alpha = 0.025
)
}
\arguments{
\item{x}{A MOBSTER fit.}

\item{bootstrap_results}{The results of running function \code{mobster_bootstrap}}

\item{bootstrap}{Type of boostrap: \code{"parametric"} or \code{"nonparametric"}}

\item{alpha}{Alpha-level empirical CIs for the bootstrap distribution.}
}
\value{
A list with the bootstrapped values, the model frequency, the bootstraped statistics and
the co-clustering probability - non-null only for nonparametric boostrap.
}
\description{
From a bootstrap run several statistics are computed.
These are the model frequency, the bootstrap statistics of each parameter
and, if the bootstrap is nonparametric, also the co-clustering probability
of the input points. All the bootstrapped values are also returned. For the
bootstrapped statistics the confidence intervals are returned for a 
required  alpha value.
}
\examples{
# Random small dataset
dataset = random_dataset(N = 200, seed = 123, Beta_variance_scaling = 100)
x = mobster_fit(dataset$data, auto_setup = 'FAST')

# Just 5 resamples of a nonparametric bootstrap run, disabling the parallel engine
options(easypar.parallel = FALSE)
boot_results = mobster_bootstrap(x$best, n.resamples = 5, auto_setup = 'FAST')

boot_stats = bootstrapped_statistics(x$best, boot_results)
print(boot_stats)
}
