% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_genes_dnds}
\alias{cancer_genes_dnds}
\title{List of cancer genes to compute dnds values.}
\format{
List of cancer genes to compute dnds values.
}
\usage{
data(cancer_genes_dnds)
}
\description{
List of gene ids (HuGO format) to compute dnds values.
There are 2 lists available of putative driver genes, and 2 lists of
essential genes; these lists have been subset to include only
genes in the RefCDS database.

Drivers have been compiled in 
`Martincorena, et al. Cell 171.5 (2017): 1029-1041`, and in 
`Tarabichi, et al. Nature Genetics 50.12 (2018): 1630`.
Essential genes have been compiled using two different cell lines in
`Wang et al. Science 350.6264 (2015): 1096-1101.` and
`Bloomen et al. Science 350.6264 (2015): 1092-1096`. All lists are 
available and named accordingly; use `names(cancer_genes_dnds)` to
see the available names.
}
\examples{
data(cancer_genes_dnds)
print(lapply(cancer_genes_dnds, head))
}
\keyword{datasets}
