% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\name{Clusters}
\alias{Clusters}
\title{Return the data with the hard clustering assigments.}
\usage{
Clusters(x, cutoff_assignment = 0)
}
\arguments{
\item{x}{A MOBSTER fit.}

\item{cutoff_assignment}{Cutoff to compute hard clustering assignments.}
}
\value{
The data stored in \code{x$data} with a column \code{label} reporting
the assigned cluster, or \code{NA} if the maximum cluster probability is below
the threshold value \code{cutoff_assignment}.
}
\description{
Extract the clustering assignments, and return the data.
Assignments can be computed subsetting the mutations that have posterior
values (reponsibilities) below a certain cutoff (default \code{0} - all 
assigments); non-assigned mutations have \code{NA} as cluster label.
}
\examples{
data('fit_example', package = 'mobster')
Clusters(fit_example$best)

# Add some cutoff to filter assignments
Clusters(fit_example$best, cutoff_assignment = .8)
}
