% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnds.R
\name{dnds}
\alias{dnds}
\title{Run a dN/dS analysis on Mobster clusters}
\usage{
dnds(x, gene_list = NULL, colors = c(Tail = "gray"), refdb = "hg19",
  dndscv_plot = "wall",
  mapping = pio:::nmfy(mobster:::.get_clusters_labels(x),
  mobster:::.get_clusters_labels(x)))
}
\arguments{
\item{x}{A MOBSTER fit object}

\item{gene_list}{An optional vector of gene names to infer dN/dS values,
default is to use the whole exome}

\item{colors}{If provided, these colours will be used for each cluster.
If a subset of colours is provided, palette Set1 from \code{RColorBrewer} is used.
By default the tail colour is provided as 'gainsboro'.}

\item{refdb}{The genome referene to use, default is to use hg19. Other references are available from
https://github.com/im3sanger/dndscv_data}

\item{dndscv_plot}{What of the dndscv scores should be visualized in a plot, by default (`wall`) is
only the global dnds value.}

\item{mapping}{The groups used to compute this statistics are defined by this variable. If
`mapping  = c(`A` = 'G1', `B` = 'G1', `C` = 'G2')`, then mutations from clusters `A` and 
`B` will be pooled into one group (`G1`), while mutations from cluster `C` will constitute
a group themselves. By default, each cluster in the fit is a own group.}
}
\value{
The fit object is a list with the summary table and the observation counts reported
by package \code{dndscv}, together with a \code{ggplot} plot.
}
\description{
This function takes a MOBSTER fit and runs dndscv (https://github.com/im3sanger/dndscv) to calculate dN/dS
values per cluster. It computes global dN/dS and per gene dN/dS values and makes a plot. dN/dS values are computed with
the best fitting MOBSTER model.
}
