% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_example}
\alias{fit_example}
\title{Example MOBSTER fit.}
\format{Data from an example MOBSTER fit.}
\usage{
data(fit_example)
}
\description{
Data from an example MOBSTER fit. This is a list returned by the fitting
algorithm, which contains the best fit, the top runs and a table summarising the fit scores.
}
\examples{
data(fit_example)
print(fit_example$best)
plot(fit_example$best)
}
\keyword{datasets}
