% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_example}
\alias{fit_example}
\title{Example MOBSTER fit.}
\format{
Output from `mobster_fit`.
}
\usage{
data(fit_example)
}
\description{
Data from an example MOBSTER fit; this object is the
result of running `mobster_fit` function on the input dat.a
The ouput is a list which contains the best fit, the top runs and 
a table summarising the fit scores. The input data has been simulated
with a stochastic branching process model.
}
\examples{
data(fit_example)
print(fit_example$best)
plot(fit_example$best)
}
\keyword{datasets}
