% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{mobster_bootstrap}
\alias{mobster_bootstrap}
\title{Bootstrap a MOBSTER fit.}
\usage{
mobster_bootstrap(x, n.resamples = 100, bootstrap = "parametric",
  cores.ratio = 0.8, cache = paste0(bootstrap, "_cache.rds"), ...)
}
\arguments{
\item{x}{An object of class \code{"dbpmm"}.}

\item{n.resamples}{Number of boostrap resamples.}

\item{bootstrap}{Type of boostrap: \code{"parametric"} or \code{"nonparametric"}}

\item{cores.ratio}{Ratio of cores to use for the parallel; see \code{?easypar}.}

\item{cache}{Cache for the computation; see \code{?easypar}.}

\item{...}{fit parameters for \code{mobster_fit}}
}
\value{
Data from the fits, resamples and a plottable figure.
}
\description{
Parmaetric and nonparametric implementation of
bootstrap estimates for MOBSTER fits. This computation is parallel
and uses  \code{?easypar}.
}
\examples{
TODO
}
