% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{mobster_bootstrap}
\alias{mobster_bootstrap}
\title{Parallel implementation of parametric/ nonparametric 
bootstrap for MOBSTER.}
\usage{
mobster_bootstrap(x, n = 100, bootstrap = "parametric", palette = "Set1",
  alpha = 1, tail.color = "darkgray", cores.ratio = 0.8, ...)
}
\arguments{
\item{x}{a fit by MOBSTER}

\item{n}{number of parametric resamples (drawn via \code{rdbpmm})}

\item{bootstrap}{"parametric" or "nonparametric"}

\item{palette}{plot palette used for Beta clusters}

\item{alpha}{transparency}

\item{tail.color}{for the tail's boxplot}

\item{cores.ratio}{ratio of cores to use}

\item{...}{fit parameters for \code{mobster_fit}}
}
\value{
Data from the fits, resamples and a plottable figure.
}
\description{
Parallel implementation of parametric/ nonparametric 
bootstrap for MOBSTER.
}
