% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_data.R
\name{mobster_dataset}
\alias{mobster_dataset}
\title{Title}
\usage{
mobster_dataset(data, samples, description = "My own dataset", na.rm = TRUE)
}
\arguments{
\item{data}{}

\item{samples}{}

\item{description}{}

\item{DP.columns}{}

\item{NV.columns}{}

\item{VAF.columns}{}
}
\description{
Title
}
