% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_2DVAF.R
\name{mobster_plt_2DVAF}
\alias{mobster_plt_2DVAF}
\title{Title}
\usage{
mobster_plt_2DVAF(obj, x, y, cluster = NULL, cluster.label = "cluster",
  marginal = FALSE, density = TRUE, cex = 1, alpha = 0.3,
  downsample = 1000, scales.fixed = TRUE, exclude = NULL,
  exclude.color = "black", marginal.remove.zeroes = T,
  palette = "Set1")
}
\arguments{
\item{marginal}{}
}
\description{
Title
}
