% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_2DVAF_matrix.R
\name{mobster_plt_2DVAF_matrix}
\alias{mobster_plt_2DVAF_matrix}
\title{Title}
\usage{
mobster_plt_2DVAF_matrix(lower.x, upper.x = NULL,
  samples = lower.x$samples, MOBSTER = !is.null(lower.x$fit.MOBSTER),
  lower.cluster = NULL, lower.cluster.label = "cluster",
  upper.cluster = NULL, upper.cluster.label = "cluster",
  title = lower.x$description, palette = list(MOBSTER = "Set1", lower =
  "Set2", upper = "Spectral"), ...)
}
\arguments{
\item{MOBSTER}{}

\item{lower.cluster}{}

\item{lower.cluster.label}{}

\item{upper.cluster}{}

\item{upper.cluster.label}{}

\item{title}{}

\item{palette}{}

\item{...}{}

\item{x}{}
}
\description{
Title
}
