% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobster_plt_2DVAF_matrix.R
\name{mobster_plt_2DVAF_matrix}
\alias{mobster_plt_2DVAF_matrix}
\title{Title}
\usage{
mobster_plt_2DVAF_matrix(x, samples = x$samples, cluster = NULL,
  cluster.label = "cluster", title = x$description, ...)
}
\arguments{
\item{x}{}

\item{samples}{}

\item{cluster}{}

\item{cluster.label}{}

\item{title}{}

\item{...}{}
}
\description{
Title
}
