% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobsterm_fit.R
\name{mobsterm_fit}
\alias{mobsterm_fit}
\title{Fit a multivariate model with MOBSTERm}
\usage{
mobsterm_fit(
  x,
  K_list = c(2, 8),
  max_iter = 1500,
  seed_list = c(123, 1234),
  lr = 0.01,
  par_threshold = 0.005,
  loss_threshold = 0.01
)
}
\arguments{
\item{x}{Input data.frame (or tibble). The input data.frame should have at least 
6 coloumns named as mutation_id, NV (number of variant), DP (depth), sample_id, purity and karyotype.}

\item{K_list}{A vector with the number of mixture components to use. All values of \code{K_list} must be positive
and strictly greater than 0.}

\item{max_iter}{Maximum number of steps for a variational inference fit. 
If parameters and loss convergence is not achieved before these steps, the fit is interrupted.}

\item{lr}{Learning rate used by the optimizer.}

\item{par_threshold}{Tolerance for parameter convergence. As ELBO oscillations are common in gradient based VI, we will monitor the convergence of all the parameters in the model,
the inference stops when (abs(new-old) / abs(old)) < par_threshold for 200 consecutive iterations, for all the parameters.}

\item{loss_threshold}{Tolerance for loss convergence. As ELBO oscillations are common in gradient based VI, we will monitor the convergence of the loss in the model,
the inference stops when (abs(new_loss-old_loss) / abs(old_loss)) < loss_threshold for 200 consecutive iterations.}

\item{seed}{A vector with the number of seeds to test for the fit.}
}
\value{
An object containing \code{best_fit} and a list of all alternative fits computed.
}
\description{
This function fits a multivariate version of the MOBSTER model implemented in \code{mobster_fit}.
}
\examples{
TBD
}
