% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_plot.R
\name{plot.dbpmm}
\alias{plot.dbpmm}
\title{Plot a MOBSTER fit.}
\usage{
plot.dbpmm(
  x,
  alpha = 0.8,
  colors = c(Tail = "gainsboro"),
  cutoff_assignment = 0,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"dbpmm"}.}

\item{alpha}{Alpha value for the colors of the histogram}

\item{colors}{If provided, these colours will be used for each cluster.
If a subset of colours is provided, palette Set1 from \code{RColorBrewer} is used.
By default the tail colour is provided as 'gainsboro'.}

\item{cutoff_assignment}{Parameters passed to run function \code{Clusters} which
returns the hard clustering assignments for the histogram plot if one wants to plot
only mutations with responsibility above this parameter.}

\item{...}{}
}
\value{
A ggplot object for the plot.
}
\description{
Plot a MOBSTER fit.
}
\examples{
data(fit_example)
plot(fit_example$best)
plot(fit_example$best, alpha = 1, cutoff_assignment = .7)
plot(fit_example$best, colors =  c(`Tail` = 'gainsboro', `C1` = 'darkorange'))
}
