% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.dbpmm}
\alias{plot.dbpmm}
\title{Plot an object of class dbpmm. This is a set of multiple frames plot via ggplot2.}
\usage{
plot.dbpmm(x, annotation = NULL, palette = "Set1",
  tail.color = "darkgray", alpha = 0.8, max.height.hist = TRUE, cex = 1,
  histogram.main = "Fit", bg.color = "ivory2", title = "MOBSTER",
  silent = FALSE, ...)
}
\arguments{
\item{x}{an object of class dbpmm}

\item{palette}{a palette for colours from RColorBrewer}

\item{tail.color}{a 2-components vector for the color of the tail, and the estimated density}

\item{alpha}{alpha channel}
}
\value{
if assigned, it return the list of ggplot2 objects.
}
\description{
Plot an object of class dbpmm. This is a set of multiple frames plot via ggplot2.
}
\examples{
something..
}
