% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobsterh_methods_plot.R
\name{plot.dbpmmh}
\alias{plot.dbpmmh}
\title{Plot a MOBSTERh fit.}
\usage{
plot.dbpmmh(x, show_na = FALSE, add_density = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{"dbpmmh"}.}

\item{show_na}{Leave in the plot points for which a clustering assignment
is not available (\code{NA}).}

\item{add_density}{Add density of the fit on top of the histogram.}

\item{...}{Unused.}
}
\value{
A ggplot or a cowplot object for the plot (depends on parameters).
}
\description{
Plot a MOBSTERh fit.
}
