% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mobsterh_methods_plot.R
\name{plot.dbpmmh}
\alias{plot.dbpmmh}
\title{Plot a MOBSTERh fit.}
\usage{
plot.dbpmmh(
  x,
  show_na = FALSE,
  add_density = TRUE,
  empty_plot = TRUE,
  assembly_plot = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"dbpmmh"}.}

\item{show_na}{Show filtered mutations as NA cluster}

\item{add_density}{Add distribution density over the histogram}

\item{empty_plot}{Pad filtered and/or absent karyotypes with empyt plots}

\item{assembly_plot}{Turn the plot into a one-line strip, instead of returning}

\item{...}{Unused.}
}
\value{
A ggplot or a cowplot object for the plot (depends on parameters).
}
\description{
Plot a MOBSTERh fit.
}
