% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boostrap_model_frequency.R
\name{plot_bootstrap_model_frequency}
\alias{plot_bootstrap_model_frequency}
\title{Plot the boostrapped model frequency}
\usage{
plot_bootstrap_model_frequency(bootstrap_results, bootstrap_statistics)
}
\arguments{
\item{bootstrap_results}{Results of a call to \code{mobster_boostrap}.}

\item{bootstrap_statisticsResults}{of a call to \code{bootstrapped_statistics}.}
}
\value{
A barplot of the boostrapped model frequency
}
\description{
From the results of a call to \code{mobster_boostrap}, and
the results of a call to \code{bootstrapped_statistics}, a barplot with
the model frequency is produced
}
\examples{
# Random small dataset
dataset = random_dataset(N = 200, seed = 123, Beta_variance_scaling = 100)
x = mobster_fit(dataset$data, auto_setup = 'FAST')

# Just 5 resamples of a nonparametric bootstrap run, disabling the parallel engine
options(easypar.parallel = FALSE)
boot_results = mobster_bootstrap(x$best, n.resamples = 5, auto_setup = 'FAST')

boot_stats = bootstrapped_statistics(x$best, boot_results)
plot_bootstrap_model_frequency(boot_results, boot_stats)
}
