% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gofit.R
\name{plot_gofit}
\alias{plot_gofit}
\title{Plot the goodness of fit.}
\usage{
plot_gofit(x, TOP = 6, nx = 3, ny = 2)
}
\arguments{
\item{x}{A list of fits computed via \code{mobster_fit}.}

\item{TOP}{The first \code{TOP} fits are used.}

\item{nx}{Columns in the matrix layout.}

\item{ny}{Rows in the matrix layout.}
}
\value{
A plot of the goodness of fit.
}
\description{
Plot the SSE (sum of squared error) as a proxy for the
goodness of fit. For the \code{TOP} available solutions the SSE trace is shown.
}
\examples{
data('fit_example', package = 'mobster')
plot_gofit(fit_example, TOP = 3)
}
