% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_init.R
\name{plot_init}
\alias{plot_init}
\title{Plot the initial density of a fit.}
\usage{
plot_init(x, colors = c(Tail = "gainsboro"))
}
\arguments{
\item{x}{An object of class \code{"dbpmm"}.}

\item{colors}{If provided, these colours will be used for each cluster.
If a subset of colours is provided, palette Set1 from \code{RColorBrewer} is used.
By default the tail colour is provided as 'gainsboro'.}
}
\value{
A ggplot object for the plot.
}
\description{
Plot the initial density of a fit.
}
\examples{
data(fit_example)
plot_init(fit_example$best)
}
