% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_variables.R
\name{plot_latent_variables}
\alias{plot_latent_variables}
\title{Plot the latent variables of the mixture.}
\usage{
plot_latent_variables(x, cutoff_assignment = 0)
}
\arguments{
\item{x}{A MOBSTER fit.}

\item{cutoff_assignment}{The parameter used to call function
\code{Clusters}, which does not assign a point to its best cluster
if the value of the corresponding latent variable is not above the cutoff.}
}
\value{
A plot of the latent variables of the mixture.
}
\description{
It renders a heatmap where the latent variables
(reponsibilities) are shown and colured according to their value.
This function also calls function \code{Clusters}, using a parameter
that determines if a point is not to be assigned its best cluster
based on a cutoff.
}
\examples{
data('fit_example', package = 'mobster')
plot_latent_variables(fit_example$best)
plot_latent_variables(fit_example$best, cutoff_assignment = .9)
}
