% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixing_proportions.R
\name{plot_mixing_proportions}
\alias{plot_mixing_proportions}
\title{Plot the mixing proportions of the mixture.}
\usage{
plot_mixing_proportions(x, colors = c(Tail = "gainsboro"))
}
\arguments{
\item{x}{An object of class 'dbpmm'.}

\item{colors}{If provided, these colours will be used for each cluster.
If a subset of colours is provided, palette Set1 from \code{RColorBrewer} is used.
By default the tail colour is provided as 'gainsboro'.}
}
\value{
A plot of the mixing proportions of the mixture.
}
\description{
Plot the mixing proportions of the mixture.
}
\examples{
data(fit_example)
plot_mixing_proportions(fit_example$best)
}
