% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_selection.R
\name{plot_model_selection}
\alias{plot_model_selection}
\title{Plot summary for model selection.}
\usage{
plot_model_selection(x, TOP = 6, nx = 3, ny = 2, ...)
}
\arguments{
\item{x}{A list of fits computed via \code{mobster_fit}.}

\item{TOP}{The first \code{TOP} fits are used (by ranking).}
}
\value{
A complex figure with all plots arranged using both
\code{ggpubr} and \code{cowplot}.
}
\description{
This plot is usefull to understand the different
fits and their rank with respect to some scoring used to
select the best model. The plot shows alternative solutions
and their rank as well.
}
\examples{
data('fit_example', package = 'mobster')
plot_model_selection(fit_example)
}
