% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{print.dbpmm}
\alias{print.dbpmm}
\title{Print a MOBSTER object.}
\usage{
\method{print}{dbpmm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'dbpmm'}.}

\item{...}{}
}
\value{
nothing.
}
\description{
Print a MOBSTER object.
}
\examples{
data(fit_example)
print(fit_example$best)
}
