% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{print.dbpmm}
\alias{print.dbpmm}
\title{Summaries for an object of class \code{'dbpmm'} is like a print.}
\usage{
\method{print}{dbpmm}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'dbpmm'}.}

\item{...}{}
}
\value{
nothing.
}
\description{
Summaries for an object of class \code{'dbpmm'} is like a print.
}
\examples{
data(fit_example)
print(fit_example$best)
}
