% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{summary.dbpmm}
\alias{summary.dbpmm}
\title{Summary for an object of class \code{'dbpmm'} is a print.}
\usage{
summary.dbpmm(object, ...)
}
\arguments{
\item{object}{An obj of class \code{'dbpmm'}.}

\item{...}{}
}
\value{
See \code{\link{print}}.
}
\description{
Summary for an object of class \code{'dbpmm'} is a print.
}
\examples{
data(fit_example)
summary(fit_example$best)
}
